fix system thread key & block num type to u64

+++ core/src/config/substrate.rs
@@ -26,7 +26,7 @@ impl Config for SubstrateConfig {
     type Address = MultiAddress<Self::AccountId, u32>;
     type Signature = MultiSignature;
     type Hasher = DynamicHasher256;
-    type Header = SubstrateHeader<u32, DynamicHasher256>;
+    type Header = SubstrateHeader<u64, DynamicHasher256>;
     type ExtrinsicParams = SubstrateExtrinsicParams<Self>;
     type AssetId = u32;
 }
@@ -369,7 +369,7 @@ mod test {
             }
         "#;
 
-        let header: SubstrateHeader<u32, BlakeTwo256> =
+        let header: SubstrateHeader<u64, BlakeTwo256> =
             serde_json::from_str(numeric_block_number_json).expect("valid block header");
         assert_eq!(header.number(), 4);
     }
@@ -389,7 +389,7 @@ mod test {
             }
         "#;
 
-        let header: SubstrateHeader<u32, BlakeTwo256> =
+        let header: SubstrateHeader<u64, BlakeTwo256> =
             serde_json::from_str(numeric_block_number_json).expect("valid block header");
         assert_eq!(header.number(), 4);
     }
